nlaugmenter/transformations/concept2sentence/requirements.txt
sibyl_tool==0.1.2